feat: `ext.rs` as arrow's extension, rm `grid.rs`
