+++ fx/src/cont/mod.rs
@@ -7,7 +7,8 @@ pub mod ab;
 pub mod batch;
 pub mod bundle;
 pub mod cvt;
-pub mod grid;
+pub mod ext;
+mod macros;
 pub mod nullopt;
 pub mod parcel;
 pub mod table;
@@ -15,7 +16,6 @@ pub mod table;
 pub use batch::*;
 pub use bundle::*;
 pub use cvt::*;
-pub use grid::*;
 pub use nullopt::*;
 pub use parcel::*;
 pub use table::*;
